Refactored the questions
In the US, Emotional Disability is a big catch-all label from the education system that can include almost any "brain" issue (literally everything from bipolar disorder to Tourette's syndrome) and also the effects of trauma/systemic oppression. In practice, it is often used as a cipher for "student's behavior disrupts the classroom." This is generally a label put on people to assign them to a special ed class, not a label that individuals use for themselves.

Referencing US law probably isn't ideal from a global standpoint as was raised in the earlier comment, so I would go ahead and hardcode some sort of definition in there. I would recommend the definition from the UN Committee on the Rights of Persons with Disabilities. (see Article 1 on page 3).

You're asking people if they have specific conditions, which is not a great idea in a global context. First, the labels are all different from country to country (for instance, cognitive, mental, and emotional disability have high overlap and are different country to country). Second, the way you've phrased the questions is asking for medical information, which is a very personal and private thing. Third, phrasing it in terms of labels is unfriendly to folks who have multiple labels or are struggling to get a diagnosis. It would be better to use the social model of disability, where we talk about environmental barriers as being the cause, and we talk about specific areas of daily life in which people are hindered by both their condition and society's choices, rather than requiring people to put a label on themselves.

Chronic illness/being sick is not included here at all. I've added a couple options that are meant to get at fatigue and chemical sensitivity and illnesses that require a lot of treatment like MS or cancer, but I'd like to have a longer think about this and maybe do some research to see how others have considered this issue.